+++ src/main/java/cz/cvut/copakond/pinkfluffyunicorn/model/items/FireItem.java
@@ -5,7 +5,7 @@ import cz.cvut.copakond.pinkfluffyunicorn.model.utils.enums.ItemEnum;
 public class FireItem extends Item {
     private static FireItem active = null; // only one fire item can be active at a time
 
-    public FireItem(int[] position, int duration) {
+    public FireItem(double[] position, int duration) {
         super("fire", position, duration, ItemEnum.FIRE);
     }
 }
