+++ src/main/java/cz/cvut/copakond/pinkfluffyunicorn/model/world/Coin.java
@@ -6,7 +6,7 @@ import cz.cvut.copakond.pinkfluffyunicorn.model.utils.enums.RenderPriorityEnums;
 public class Coin extends GameObject {
     static int coinsLeft = 0;
 
-    public Coin(int[] position) {
+    public Coin(double[] position) {
         super("coin", position, RenderPriorityEnums.ITEM.getValue());
     }
