+++ src/main/java/cz/cvut/copakond/pinkfluffyunicorn/model/world/Tile.java
@@ -5,7 +5,7 @@ import cz.cvut.copakond.pinkfluffyunicorn.model.utils.enums.RenderPriorityEnums;
 
 public class Tile extends GameObject {
     int textureType;
-    public Tile(int[] position, int textureType) {
+    public Tile(double[] position, int textureType) {
         super("tile", position, RenderPriorityEnums.TILE.getValue());
         this.textureType = textureType;
         super.loadTextures("tile", new int[]{textureType});
